requirements.txt
boto3==1.26.51
fastapi==0.89.1
nltk==3.7
numpy==1.23.0
pandas==1.4.3
Pillow==9.4.0
requests==2.28.1
torch==1.11.0
torchvision==0.12.0
transformers==4.20.1
uvicorn==0.18.2